Update Portal.yaml
I noticed Data Series and Data Catalog are being populated with NF-OSI Processed Data and the data catalog URL, in cases where it is not applicable.  I'm wondering if this is because there is only a single permissible value?  Added another value (Not Applicable) . 


Also fixed some minor formatting issues.