Update Portal.yaml (#690)
* Update Portal.yaml

I noticed Data Series and Data Catalog are being populated with NF-OSI Processed Data and the data catalog URL, in cases where it is not applicable.  I'm wondering if this is because there is only a single permissible value?  Added another value (Not Applicable) . 


Also fixed some minor formatting issues.

* Rebuild NF.jsonld, json

* Update Portal.yaml

* Merge from main

* Rebuild NF.jsonld, json

---------

Co-authored-by: nf-osi[bot] <nf-osi@sagebionetworks.org>
Co-authored-by: Anh Nguyet Vu <32753274+anngvu@users.noreply.github.com>
Co-authored-by: Anh Nguyet Vu <anngvu@gmail.com>